product type initializer added

+++ src/main/java/com/GraduationProject/GraduationProject/entity/ProductType.java
@@ -18,7 +18,7 @@ public class ProductType {
     private String productName;
 
     @Column(nullable = false)
-    private Float effect;
+    private Double effect;
 
 
 }
